ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cerebras-cloud-sdk>=1.29.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
]
